# backend/src/common/types/prompt.types.ts
/**
 * Prompt Types
 *
 * Defines text-to-video prompt structures and moderation status.
 */

/**
 * Content moderation status
 */
export enum ModerationStatus {
  PENDING = 'pending',
  APPROVED = 'approved',
  FLAGGED = 'flagged',
  BYPASSED = 'bypassed', // User explicitly approved despite flags
}

/**
 * GenerationPrompt represents the text-to-video prompt for Sora 2
 */
export interface GenerationPrompt {
  /** Unique identifier (UUID) */
  promptId: string;

  /** AI-generated prompt text */
  generatedText: string;

  /** User's edited version (if modified, optional) */
  userEditedText?: string;

  /** Final approved text (generated or edited) */
  finalText: string;

  /** Character count of final text */
  characterCount: number;

  /** Timestamp when generated by AI */
  generatedAt: Date;

  /** Timestamp when approved by user (optional) */
  approvedAt?: Date;

  /** Content moderation result */
  moderationStatus: ModerationStatus;

  /** Flagged content categories (if any, optional) */
  moderationFlags?: string[];
}