# specs/001-ugc-video-generator/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Viral2Viral - UGC Video Cloner API
  description: |
    API for generating advertisement videos based on UGC (User Generated Content) analysis.
    
    ## Workflow
    1. Create session
    2. Upload source video to S3
    3. Trigger video analysis (Gemini AI)
    4. Submit product information
    5. Upload product image to S3
    6. Generate video prompt (GPT-5)
    7. Generate video (Sora 2)
    8. Download generated video
    
    This is a stateless POC - session data is held in memory only.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /sessions/{sessionId}:
    get:
      tags:
        - Session
      summary: Get session details
      description: Retrieve complete session state including all workflow data
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSessionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /sessions/{sessionId}/generate:
    post:
      tags:
        - Generation
      summary: Generate video
      description: Start video generation using Sora 2 with approved prompt and product image
      operationId: generateVideo
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '202':
          description: Video generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateVideoResponse'
        '400':
          description: Missing required data (prompt not approved or image not uploaded)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Generation
      summary: Get video generation status
      description: Poll for video generation status and retrieve download URL when complete
      operationId: getVideoStatus
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Generation status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVideoStatusResponse'
        '404':
          description: Generation not started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Unique session identifier (UUID)
      schema:
        type: string
        format: uuid
        example: 123e4567-e89b-12d3-a456-426614174000
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      required:
        - timestamp
        - requestId
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - error
          properties:
            error:
              $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VIDEO_TOO_LARGE
        message:
          type: string
          example: Video file exceeds maximum size of 100MB

    # Session schemas
...
    GetSessionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/Session'
    Session:
      type: object
      required:
        - sessionId
        - status
        - createdAt
        - lastActivityAt
      properties:
        sessionId:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, video_uploaded, analyzing, analysis_complete, product_info_added, prompt_generated, generating_video, video_complete, error]
        createdAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        originalVideo:
          $ref: '#/components/schemas/OriginalVideo'
        videoAnalysis:
          $ref: '#/components/schemas/VideoAnalysis'
        productInformation:
          $ref: '#/components/schemas/ProductInformation'
        generationPrompt:
          $ref: '#/components/schemas/GenerationPrompt'
        generatedVideo:
          $ref: '#/components/schemas/GeneratedVideo'

    # Video schemas
...
    OriginalVideo:
      type: object
      required:
        - s3Key
        - s3Bucket
        - fileName
        - fileSize
        - mimeType
        - uploadedAt
      properties:
        s3Key:
          type: string
        s3Bucket:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        mimeType:
          type: string
        uploadedAt:
          type: string
          format: date-time
        duration:
          type: number
          format: float
          description: Video duration in seconds
        thumbnailS3Key:
          type: string
        downloadUrl:
          type: string
          format: uri
          description: Presigned download URL (temporary)

    # Analysis schemas
...
    VideoAnalysis:
      type: object
      required:
        - analysisId
        - analyzedAt
        - status
        - sceneBreakdown
      properties:
        analysisId:
          type: string
          format: uuid
        analyzedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, processing, complete, failed]
        sceneBreakdown:
          type: string
          description: Raw scene-by-scene breakdown
        userEdits:
          type: string
          description: User's edited version
        error:
          $ref: '#/components/schemas/AnalysisError'
    AnalysisError:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    # Product schemas
...
    ProductInformation:
      type: object
      required:
        - productName
        - productDescription
        - addedAt
      properties:
        productName:
          type: string
        productDescription:
          type: string
        productImageS3Key:
          type: string
        productImageMimeType:
          type: string
        addedAt:
          type: string
          format: date-time
        downloadUrl:
          type: string
          format: uri

    # Prompt schemas
...
    GenerationPrompt:
      type: object
      required:
        - promptId
        - generatedText
        - finalText
        - characterCount
        - generatedAt
        - moderationStatus
      properties:
        promptId:
          type: string
          format: uuid
        generatedText:
          type: string
          description: AI-generated prompt
        userEditedText:
          type: string
          description: User's edited version
        finalText:
          type: string
          description: Final approved text
        characterCount:
          type: integer
        generatedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        moderationStatus:
          type: string
          enum: [pending, approved, flagged, bypassed]
        moderationFlags:
          type: array
          items:
            type: string

    # Generation schemas
    GenerateVideoResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/GeneratedVideo'
    GetVideoStatusResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/GeneratedVideo'
    GeneratedVideo:
      type: object
      required:
        - generatedVideoId
        - s3Key
        - s3Bucket
        - fileName
        - mimeType
        - status
        - initiatedAt
      properties:
        generatedVideoId:
          type: string
          format: uuid
        s3Key:
          type: string
        s3Bucket:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        mimeType:
          type: string
        status:
          type: string
          enum: [pending, processing, complete, failed]
        initiatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        estimatedCompletionTime:
          type: string
          format: date-time
        downloadUrl:
          type: string
          format: uri
        error:
          $ref: '#/components/schemas/GenerationError'
    GenerationError:
      type: object
      required:
        - code
        - message
        - timestamp
        - retryable
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        retryable:
          type: boolean
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional for POC)

# Security is optional for POC but documented for future use
security: []